requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
aio-pika
aiosqlite
alembic
asyncpg
sqlalchemy
pytest
pytest-asyncio
python-json-logger
PyYAML